完成了密码加密与微服务鉴权JWT

+++ tensquare_user/src/main/java/com/tensquare/user/service/UserService.java
@@ -9,7 +9,9 @@ import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 import javax.persistence.criteria.Selection;
+import javax.servlet.http.HttpServletRequest;
 
+import io.jsonwebtoken.Claims;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,6 +27,7 @@ import util.IdWorker;
 
 import com.tensquare.user.dao.UserDao;
 import com.tensquare.user.pojo.User;
+import util.JwtUtil;
 
 /**
  * 服务层
@@ -49,6 +52,13 @@ public class UserService {
 
 	@Autowired
 	private BCryptPasswordEncoder encoder;
+
+	@Autowired
+	private JwtUtil jwtUtil;
+
+	@Autowired
+	private HttpServletRequest request;
+
 	/**
 	 * 查询全部列表
 	 * @return
@@ -117,10 +127,33 @@ public class UserService {
 	}
 
 	/**
-	 * 删除
+	 *
+	 * 删除:必须拥有admin角色的人才可以删除user
 	 * @param id
 	 */
 	public void deleteById(String id) {
+		/*String head = request.getHeader("Authorization");
+		if(head == null || head.equals("")){
+			throw new RuntimeException("权限不足");
+		}
+		if(!head.startsWith("Bearer ")){
+			throw new RuntimeException("权限不足");
+		}
+		String token = head.substring(7);
+
+		try {
+			Claims claims = jwtUtil.parseJWT(token);
+			String roles = (String) claims.get("roles");
+			if(roles==null || !roles.equals("admin")){
+				throw new RuntimeException("权限不足");
+			}
+		} catch (Exception e) {
+			throw new RuntimeException("权限不足");
+		}*/
+		String token = (String) request.getAttribute("claim_admin");
+		if(token ==null || "".equals(token)){
+			throw new RuntimeException("权限不足");
+		}
 		userDao.deleteById(id);
 	}
